Stop passing around options
Fixes issue in the previous commit where the defaults wouldn't be
applied when the user supplied a "valid" config.

We used to pass options around and have assorted checks throughout the
code to set defaults. Since we want to have defaults set by the time we
leave the defaults.rs file, we shouldn't be passing around options.

With one exception:
Since we have to deal with options for white/blacklist, that is kept as
an option.