Remove word after it is typed correctly
By doing this, we can ensure users can type out whole passages no matter
their terminal size.

This will be added as a config option, but may become the default.

Note that there are zero tests for this, and it is not configurable. It
could be a little cleaner too. Do NOT merge yet.

Full text should also show at the end on complete rather than the last
word.

Commit also cleans up a little bit of duplicate logic we had for game logic
that actually was never being run.