Properly terminate in all cases
Previously we would not correctly terminate in instant death mode. This
did not noticeably manifest because the only difference was that we
would continue showing more errors as they happened, and C-u would make
a small graphical bug.

This also uses the proper "complete" variable that we set when a passage
is complete rather than the brittle - check is the idx is the same as
the length of the words, as we may terminate in cases that are not just
completing the passage.

Fixes: #23