The splittening of modules
game/mod.rs has been split to 2 modules. Previously it was over 500
lines which is just insane.

formatter was created as a text formatter for the game module.

indexer was created as the index resolver for the game module.

Some of the methods in these fellas could probably be broken up a little
more, but this will suffice for now.