Add repo check and UI for extra repos
This adds the rest of the work for checking what version extra repos
are, and cloning them in.

This also shows error messages, and ensures typeracer does not explode
if it hits an error along the way. It neatly shows the error to the user
(something that really should have happened a long time ago).