Ensure up and down keys actually do something different in graphs
Since there were only two states, up and down could do the same thing.
Now that there are 3, we'll want to make sure we can go forward and
backwards.

This PR adds a simple "prev" impl to move backwards the same way we move
forward.

We may want to consider using enum ordinalize in the future as an
alternative to this, as it is beginning to get unwieldy and errorprone.

https://docs.rs/enum-ordinalize/3.1.8/enum_ordinalize/