Fixup lang-pack rendering
During a previous tui update, some of the semantics changed and
new-lines would no longer be rendered in spans the same way they were
before. Now in order to properly render newlines, we need to make a Text
object and extend it for each of the instructions.

Now rather than looking like a collapsed single sentence, the lang pack
downloader shows everything in a set of lines again.