Migrate to async input for all of the other modules that matter
We don't actually need async input for the lang pack or config portion
of the game.

Well maybe we could use it for the lang pack downloading portion, but
really we would want callbacks and all that for the downloads too before
we do that. Not something I care about enough right now.

This should already be quite an upgrade.