fixed bug where very fast input was dropped
Previously, we were calling 'stdin.keys()' on every iteration of the
input loop, then fetching the next key input and discarding the 'Keys'
iterator and reconstructing it on the next iteration. The problem with
this is that 'Keys' has an internal buffer, so if the user entered input
sufficiently quickly, it would be dropped between iterations.

Instead, we are now constructing a single 'Keys' instance at the start
of the loop, so the internal buffer should persist.