Provide patch for handling old db installations
Regretfully this is an old bug that has hung around for far too long
without a proper solution. The only remedy for now appears to be to
delete the database and restart. Might as well bake this into the code
so folks don't have to keep doing this on their own.

Safeguards are in place to ensure we are only deleting old databases.

fixes: https://gitlab.com/ttyperacer/terminal-typeracer/-/issues/39